+++ Sources/Vapor/Request/RequestContent.swift
@@ -11,108 +11,6 @@ extension String: RequestContentSubscript { }
 extension Int: RequestContentSubscript {}
 
 public extension Request {
-
-    public enum MultiPart: Node {
-        case files([(name: String?, type: MediaType?, data: Data)])
-        case file(name: String?, type: MediaType?, data: Data)
-        case input(String)
-        case inputArray([String])
-        
-        public var file: (name: String?, type: MediaType?, data: Data)? {
-            if case .file(let name, let media, let data) = self {
-                return (name, media, data)
-            }
-
-            return nil
-        }
-
-        public var files: [(name: String?, type: MediaType?, data: Data)]? {
-            if case .files(let files) = self {
-                return files
-            }
-
-            return nil
-        }
-
-        public var input: String? {
-            if case .input(let string) = self {
-                return string
-            }
-
-            return nil
-        }
-
-        public var inputArray: [String]? {
-            if case .inputArray(let array) = self {
-                return array
-            }
-            
-            return nil
-        }
-        
-        public var isNull: Bool {
-            return self.input == "null"
-        }
-        
-        public var bool: Bool? {
-            if case .input(let bool) = self {
-                return Bool(bool)
-            }
-            
-            return nil
-        }
-        
-        public var int: Int? {
-            guard let double = double else { return nil }
-            return Int(double)
-        }
-        
-        public var uint: UInt? {
-            guard let double = double else { return nil }
-            return UInt(double)
-        }
-        
-        public var float: Float? {
-            guard let double = double else { return nil }
-            return Float(double)
-        }
-        
-        public var double: Double? {
-            if case .input(let d) = self {
-                return Double(d)
-            }
-            
-            return nil
-        }
-        
-        public var string: String? {
-            return self.input
-        }
-        
-        public var array: [Node]? {
-            guard case .input(let a) = self else {
-                return nil
-            }
-            
-            return [a]
-//            return self
-//                .split(byString: ",")
-//                .map { $0 as Node }
-        }
-        
-        public var object: [String : Node]? {
-            return nil
-        }
-        
-        public var json: Json? {
-            if case .input(let j) = self {
-                return Json(j)
-            }
-            
-            return nil
-        }
-    }
-
     /**
         The data received from the request in json body or url query
     */
